Make BufferDesc::width UINT64

+++ Sorcery/src/rendering/graphics.cpp
@@ -783,7 +783,7 @@ auto GraphicsDevice::CreateBufferViews(ID3D12Resource2& buffer, BufferDesc const
                                        UINT& uav) -> void {
   if (desc.constant_buffer) {
     cbv = res_desc_heap_.Allocate();
-    D3D12_CONSTANT_BUFFER_VIEW_DESC const cbv_desc{buffer.GetGPUVirtualAddress(), desc.size};
+    D3D12_CONSTANT_BUFFER_VIEW_DESC const cbv_desc{buffer.GetGPUVirtualAddress(), static_cast<UINT>(desc.size)};
     device_->CreateConstantBufferView(&cbv_desc, res_desc_heap_.GetDescriptorCpuHandle(cbv));
   } else {
     cbv = details::kInvalidResourceIndex;
@@ -794,7 +794,7 @@ auto GraphicsDevice::CreateBufferViews(ID3D12Resource2& buffer, BufferDesc const
     D3D12_SHADER_RESOURCE_VIEW_DESC const srv_desc{
       .Format = DXGI_FORMAT_UNKNOWN, .ViewDimension = D3D12_SRV_DIMENSION_BUFFER,
       .Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING,
-      .Buffer = {0, desc.size / desc.stride, desc.stride, D3D12_BUFFER_SRV_FLAG_NONE}
+      .Buffer = {0, static_cast<UINT>(desc.size / desc.stride), desc.stride, D3D12_BUFFER_SRV_FLAG_NONE}
     };
     device_->CreateShaderResourceView(&buffer, &srv_desc, res_desc_heap_.GetDescriptorCpuHandle(srv));
   } else {
@@ -806,8 +806,8 @@ auto GraphicsDevice::CreateBufferViews(ID3D12Resource2& buffer, BufferDesc const
     D3D12_UNORDERED_ACCESS_VIEW_DESC const uav_desc{
       .Format = DXGI_FORMAT_UNKNOWN, .ViewDimension = D3D12_UAV_DIMENSION_BUFFER,
       .Buffer = {
-        .FirstElement = 0, .NumElements = desc.size / desc.stride, .StructureByteStride = desc.stride,
-        .CounterOffsetInBytes = 0, .Flags = D3D12_BUFFER_UAV_FLAG_NONE
+        .FirstElement = 0, .NumElements = static_cast<UINT>(desc.size / desc.stride),
+        .StructureByteStride = desc.stride, .CounterOffsetInBytes = 0, .Flags = D3D12_BUFFER_UAV_FLAG_NONE
       }
     };
     device_->CreateUnorderedAccessView(&buffer, nullptr, &uav_desc, res_desc_heap_.GetDescriptorCpuHandle(uav));
